qkc/common: move token balances from qkc/types and add MNT DefaultTokenID (#42)
* qkc/common: move token balances from qkc/types

Move TokenBalances and its tests from qkc/types to qkc/common with no
logic change (package rename only). A follow-up commit adds the MNT
account-encoding helpers on top.


* qkc/common: add DefaultTokenID and NewTokenBalancesFromBytes for MNT

Add the helpers the MNT account-encoding layer depends on, on top of the
moved token balances:

- DefaultTokenID (= TokenIDEncode("QKC") = 35760): the QKC native token
  ID carried inside the multi-token balance set when merging into the
  QuarkChain account trie encoding.
- NewTokenBalancesFromBytes: alias for NewTokenBalances, kept so MNT
  call sites can use the more descriptive name without churn.
- Expand the TokenTrieThreshold doc comment explaining why the >16-token
  trie format is intentionally not implemented in this package.
- Tests for the two new symbols.

Zero-value handling is unchanged from the moved implementation.


* Delete alias function NewTokenBalancesFromBytes

Removed alias function NewTokenBalancesFromBytes.

---------
