+++ qkc/common/token_test.go
@@ -2,7 +2,7 @@
 
 // Token balance tests exercise pyquarkchain-compatible QKC wire bytes.
 
-package types
+package common
 
 import (
 	"encoding/hex"
@@ -27,6 +27,12 @@ func testU256Decimal(t *testing.T, v string) *uint256.Int {
 		t.Fatal(err)
 	}
 	return data
+}
+
+func TestDefaultTokenID(t *testing.T) {
+	// DefaultTokenID must equal TokenIDEncode("QKC") = 35760.
+	assert.Equal(t, uint64(35760), DefaultTokenID)
+	assert.Equal(t, TokenIDEncode("QKC"), DefaultTokenID)
 }
 
 func TestNewTokenBalanceMap(t *testing.T) {
